Move Skylark to its own package
