+++ src/main/java/skylark/command/Command.java
@@ -1,6 +1,6 @@
 package skylark.command;
 
-import skylark.main.SkylarkException;
+import skylark.skylark.SkylarkException;
 import skylark.task.Deadline;
 import skylark.task.Event;
 import skylark.task.Task;
