+++ src/main/java/skylark/main/Main.java
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import skylark.skylark.Skylark;
 import skylark.ui.MainWindow;
 
 /**
