+++ src/main/java/skylark/task/Deadline.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
-import skylark.main.SkylarkException;
+import skylark.skylark.SkylarkException;
 
 /** Represents a Deadline Task. */
 public class Deadline extends Task {
