+++ src/main/java/skylark/ui/MainWindow.java
@@ -7,7 +7,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
-import skylark.main.Skylark;
+import skylark.skylark.Skylark;
 
 /**
  * Controller for MainWindow. Provides the layout for the other controls.
